changing float to double
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51018 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/awt/font/GTextLayout.java
@@ -11,7 +11,7 @@ public interface GTextLayout {
 	/**
 	 * @return width of the text in pixels
 	 */
-	float getAdvance();
+	double getAdvance();
 
 	/**
 	 * @return bounding rectangle
@@ -22,7 +22,7 @@ public interface GTextLayout {
 	 * 
 	 * @return distance between baseline and top of the highest character
 	 */
-	float getAscent();
+	double getAscent();
 
 	/**
 	 * Draw this text in graphics
@@ -39,6 +39,6 @@ public interface GTextLayout {
 	/**
 	 * @return distance between baseline and deepest bottom of a character
 	 */
-	float getDescent();
+	double getDescent();
 
 }
